orchestrator/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pyyaml==6.0.2
psutil==6.1.0
requests==2.32.3
rich==13.9.2
